requirements.txt
keyboard
pynput
pytesseract
scikit-image
pillow